command issue not done yet

+++ src/exec/here_doc.c
@@ -6,19 +6,20 @@
 /*   By: yishan <yishan@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/09 12:58:45 by yishan            #+#    #+#             */
-/*   Updated: 2025/04/09 22:24:08 by yishan           ###   ########.fr       */
+/*   Updated: 2025/04/13 12:28:26 by yishan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/shell_data.h"
 #include "../../libft/inc/libft.h"
 
-/*static bool	read_heredoc_input(t_data *data, int fd, char *delimiter)
+static t_bool	read_heredoc_input(t_data *data, int fd, char *delimiter)
 {
 	char	*line;
 
-	while (true)
+	while (1)
 	{
+		line = NULL;
 		line = readline("> ");
 		if (!line)
 		{
@@ -33,18 +34,18 @@
 			free(line);
 			break ;
 		}
-		if (!replace_dollar(&line, data))
+		if (!dollar_handle(&line, data))
 		{
 			free(line);
 			//free_all(data, ERR_MALLOC, EXT_MALLOC);
-			return (false);
+			return (FALSE);
 		}
 		ft_putstr_fd(line, fd);
 		ft_putchar_fd('\n', fd);
 		free(line);
 	}
-	return (true);
-}*/
+	return (TRUE);
+}
 
 int	here_doc(t_data *data, char *delimiter)
 {
@@ -53,14 +54,13 @@ int	here_doc(t_data *data, char *delimiter)
 	fd = open(".heredoc.tmp", O_CREAT | O_WRONLY | O_TRUNC, 0644);
 	if (fd < 0)
 		return (-1);
-	/*if (!read_heredoc_input(data, fd, delimiter))
+	if (!read_heredoc_input(data, fd, delimiter))
 	{
 		close(fd);
 		unlink(".heredoc.tmp");
 		return (-1);
 	}
-	close(fd)*/;
-
+	close(fd);
 	fd = open(".heredoc.tmp", O_RDONLY);
 	if (fd >= 0)
 		unlink(".heredoc.tmp");
